Add files via upload
Problem 1: (Infinite Algorithm) The problem with this algorithm is that, some of the steps appear more than once, i.e. step 5 get second number, step 7, get third number, etc.

Problem 2: (Finite Algorithm) The new list of numbers is given as 26, 49, 498, 9387, 48962, 1, -1. The value –1 is a unique number since all other numbers are positive. This means that the procedure will stop once -1 is encountered.